requirements.txt
Django==3.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-model-utils==4.0.0
PyJWT==v1.7.1
model-bakery==1.7.*
